refactor: 클래스분리
사용자 입력 파일
결과값 출력 파일
